Build: Fix #2228, by correcting the type used.

+++ glslang/MachineIndependent/ParseHelper.cpp
@@ -8527,7 +8527,7 @@ const TTypeList* TParseContext::recordStructCopy(TStructRecord& record, const TT
     size_t memberCount = tmpType->getStruct()->size();
     size_t originHash = 0, tmpHash = 0;
     std::hash<size_t> hasher;
-    for (uint32_t i = 0; i < memberCount; i++) {
+    for (size_t i = 0; i < memberCount; i++) {
         size_t originMemberHash = hasher(originType->getStruct()->at(i).type->getQualifier().layoutPacking +
                                          originType->getStruct()->at(i).type->getQualifier().layoutMatrix);
         size_t tmpMemberHash = hasher(tmpType->getStruct()->at(i).type->getQualifier().layoutPacking +
